requirements.txt
acres==0.5.0
aistudio-sdk==0.3.8
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
bce-python-sdk==0.9.49
bcrypt==5.0.0
billiard==4.2.2
blinker==1.9.0
boto3==1.40.67
botocore==1.40.67
cachetools==6.2.1
captcha==0.7.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.10.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
Django==5.2.8
django-environ==0.12.0
django_celery_results==2.6.0
djangorestframework==3.16.1
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
extra-streamlit-components==0.1.81
filelock==3.20.0
fitz==0.0.1.dev2
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
importlib_resources==6.5.2
intel-openmp==2021.4.0
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl==2021.4.0
modelscope==1.31.0
mpmath==1.3.0
narwhals==2.10.1
networkx==3.4.2
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==1.26.4
oauth2client==4.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
polars==1.35.1
polars-runtime-32==1.35.1
prettytable==3.16.0
prompt_toolkit==3.0.52
protobuf==6.33.0
prov==2.1.1
psutil==7.1.3
puremagic==1.30
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
PyJWT==2.10.1
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdfium2==5.0.0
Pyrebase4==4.8.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.2
rdflib==7.4.0
redis==7.0.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==0.10.1
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.13.1
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.51.0
streamlit-authenticator==0.4.2
sympy==1.14.0
tbb==2021.13.1
tenacity==9.1.2
tifffile==2025.5.10
tokenizers==0.19.1
toml==0.10.2
torch==2.9.0+cu126
torchaudio==2.9.0+cu126
torchvision==0.24.0+cu126
tornado==6.5.2
tqdm==4.67.1
traits==7.0.2
transformers==4.40.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.225
ultralytics-thop==2.0.18
urllib3==1.26.20
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
xlsxwriter==3.2.9
